drivers: i2s: add native_sim file backend
Add a native_sim file-backed I2S controller for RX, TX,
and loopback tests.

The driver reads and writes PCM data through host-libc
file I/O and supports configurable stream direction.
It exposes standard I2S trigger handling, provides
command-line overrides for the RX and TX file paths,
is documented in the native_sim board guide, and adds
I2S to the supported peripherals in native_sim.yaml.

+++ samples/drivers/i2s/echo/src/main.c
@@ -340,7 +340,7 @@ int main(void)
 
 		while (k_sem_take(&toggle_transfer, K_NO_WAIT) != 0) {
 			void *mem_block;
-			uint32_t block_size;
+			size_t block_size;
 			int ret;
 
 			ret = i2s_read(i2s_dev_rx, &mem_block, &block_size);
